[MOV] base: split partner sub models
RATIONALE

Soon contact fields of some main person-related models in Odoo will be
updated to try to use an identity concept. Some code cleaning is performed
to ease review of future changes.

PURPOSE

Purpose of this commit is to split models to have a file per model. Partner
model in base still holds code for its sub models: tags, title, industry
and the famous format mixin. Those now lie in their own file to ease
model understanding when looking at file organization.

LINKS

Task ID-2477479 (identity task)
COM PR odoo/odoo# 67395
ENT PR odoo/enterprise# 17617
UPG PR odoo/upgrade# 2382
