+++ odoo/addons/base/models/__init__.py
@@ -33,11 +33,15 @@ from . import ir_demo_failure
 from . import report_layout
 from . import report_paperformat
 
+from . import format_address_mixin
 from . import image_mixin
 
 from . import res_country
 from . import res_lang
 from . import res_partner
+from . import res_partner_category
+from . import res_partner_industry
+from . import res_partner_title
 from . import res_bank
 from . import res_config
 from . import res_currency
