+++ odoo/addons/base/models/format_address_mixin.py
@@ -1,0 +1,33 @@
+# -*- coding: utf-8 -*-
+# Part of Odoo. See LICENSE file for full copyright and licensing details.
+
+from lxml import etree
+
+from odoo import models
+
+
+class FormatAddressMixin(models.AbstractModel):
+    _name = "format.address.mixin"
+    _description = 'Address Format'
+
+    def _fields_view_get_address(self, arch):
+        # consider the country of the user, not the country of the partner we want to display
+        address_view_id = self.env.company.country_id.address_view_id.sudo()
+        if address_view_id and not self._context.get('no_address_format') and (not address_view_id.model or address_view_id.model == self._name):
+            #render the partner address accordingly to address_view_id
+            doc = etree.fromstring(arch)
+            for address_node in doc.xpath("//div[hasclass('o_address_format')]"):
+                Partner = self.env['res.partner'].with_context(no_address_format=True)
+                sub_view = Partner.fields_view_get(
+                    view_id=address_view_id.id, view_type='form', toolbar=False, submenu=False)
+                sub_view_node = etree.fromstring(sub_view['arch'])
+                #if the model is different than res.partner, there are chances that the view won't work
+                #(e.g fields not present on the model). In that case we just return arch
+                if self._name != 'res.partner':
+                    try:
+                        self.env['ir.ui.view'].postprocess_and_fields(sub_view_node, model=self._name)
+                    except ValueError:
+                        return arch
+                address_node.getparent().replace(address_node, sub_view_node)
+            arch = etree.tostring(doc, encoding='unicode')
+        return arch
